requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic==2.5.0

# LLM and processing
openai==1.3.7
tiktoken==0.5.1

# LaTeX processing
pylatexenc==2.10

# Async processing
celery==5.3.4
redis==5.0.1

# File handling
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
